StockFlow

StockFlow.TheoryStockAndFlow0 — Constant

define the sub-schema of c0, which includes the three objects: stocks(S), sum-auxiliary-variables(SV), and the linkages between them (LS) to be composed

StockFlow.TheoryStockAndFlowF — Constant

define the schema of a general stock and flow diagram

StockFlow.TheoryStockAndFlowStructure — Constant

define the schema of a general stock and flow diagram, not includes the functions of auxiliary variables

StockFlow.TheoryStockAndFlowStructureF — Constant

define the schema of a general stock and flow diagram

StockFlow.CausalLoop — Method

CausalLoop(ns,es)

Create causal loop diagram from collection of nodes and collection of edges.

StockFlow.StockAndFlow0 — Method

StockAndFlow0(s,sv,ssv)

for an instance of the sub-schema, the program supports only have stocks, or only have sum auxiliary variables, or both stocks and sum auxiliary variables, or have both

StockFlow.StockAndFlowF — Method

StockAndFlowF(s,p,v,f,sv)

EXAMPLE: when define the dynamical variables, need to define with the correct order

sir_StockAndFlow=StockAndFlowF((:S=>(:F_NONE,:inf,:N), :I=>(:inf,:F_NONE,:N)),
 (:c, :beta),
 (:v_prevalence=>(:I,:N,:/),:v_meanInfectiousContactsPerS=>(:c,:v_prevalence,:*),:v_perSIncidenceRate=>(:beta,:v_meanInfectiousContactsPerS,:*),:v_newInfetions=>(:S,:v_perSIncidenceRate,:*)),
 (:inf=>:v_newInfetions),
 (:N))

StockFlow.add_prefix! — Method

add_prefix!(sf::AbstractStockAndFlowStructureF, prefix)

Modify a AbstractStockAndFlowStructureF so named elements begin with prefix Prefix can be anything which can be cast to a Symbol.For feet.

StockFlow.add_prefix! — Method

add_prefix!(sf::AbstractStockAndFlowStructureF, prefix)

Modify a AbstractStockAndFlowStructureF so named elements begin with prefix Prefix can be anything which can be cast to a Symbol.

StockFlow.add_suffix! — Method

add_suffix!(sf::AbstractStockAndFlow0, suffix)

Modify a AbstractStockAndFlow0 so named elements end with suffix. Suffix can be anything which can be cast to a Symbol. For feet.

StockFlow.add_suffix! — Method

add_suffix!(sf::AbstractStockAndFlow0, suffix)

Modify a AbstractStockAndFlow0 so named elements end with suffix. Suffix can be anything which can be cast to a Symbol.

StockFlow.args — Method

args(p::AbstractStockAndFlowF,v)

Return a Vector of Symbols of flattened stocks, sums, parameters and source dynamic variables a dynamic variable at index v links to.

StockFlow.args — Method

args(p::AbstractStockAndFlowStructureF,v)

Return a Vector of Symbols of flattened stocks, sums, parameters and source dynamic variables a dynamic variable at index v links to.

StockFlow.args_vname — Method

Return a Tuple of Vectors of Symbols of flattened stocks, sums, parameters and source dynamic variables a dynamic variable at index v links to.

StockFlow.convertStockFlowToSystemStructure — Method

Return a new StockAndFlowStructureF with flattened names, operators and positions from an AbstractStockAndFlowF.

StockFlow.convertSystemStructureToStockFlow — Method

Return a new stock flow with flattened names, operators and positions from an AbstractStockAndFlowStructureF. Need to provide dynamic variable definitions, eg

convertSystemStructureToStockFlow(MyStockFlowStructure, (:v_prevalence=>(:I,:N,:/),:v_meanInfectiousContactsPerS=>(:c,:v_prevalence,:*)))

StockFlow.convertToCausalLoop — Method

Convert StockFlow to CLD. Nodes: stocks, flows, sum variables, parameters, nonflow dynamic variables Edges: morphisms in stock flow

StockFlow.extracFlowsStructureAndFlatten — Method

Return flow names as Symbol, along with the linked flow variables

StockFlow.extracPsStructureAndFlatten — Method

Return parameter names as Symbol

StockFlow.extracStocksStructureAndFlatten — Method

Return stock names as Symbol, along with the linked flows and sum variables

StockFlow.extracSumVStructureAndFlatten — Method

Return sum variable names as Symbol, along with the linked dynamic variables

StockFlow.extracVAndAttrStructureAndFlatten — Method

Convert dynamic variable names to Symbol, convert all operators to a single operator if they are equal else throw an error, and

StockFlow.extracVStructureAndFlatten — Method

Return dynamic variable definitions as Vector with elements of form :dv => [:arg1, :arg2]

StockFlow.flowVariableIndex — Method

return the auxiliary variable's index that related to the flow with index of f

StockFlow.fname — Method

return the flows name with index of f

StockFlow.fnames — Method

return flow names

StockFlow.funcDynam — Method

funcDynam(p::AbstractStockAndFlow,v)

return the functions of variables give index v

StockFlow.funcDynam — Method

funcDynam(sf::AbstractStockAndFlowF,v)

return the functions of variables give index v

StockFlow.funcFlow — Method

generate the function substituting sum variables in with flow index fn

StockFlow.funcFlowRaw — Method

return the functions (not substitutes the function of sum variables yet) of flow index f

StockFlow.funcFlows — Method

return the LVector of pairs: fname => function (with function of sum variables substitue in)

StockFlow.funcFlowsRaw — Method

return the LVector of pairs: fname => function (raw: not substitutes the function of sum variables yet)

StockFlow.funcSV — Method

generate the function of a sum auxiliary variable (index sv) with the sum of all stocks links to it

StockFlow.funcSVs — Method

return the LVector of pairs: svname => function

StockFlow.fv — Method

get flow variable of flow

StockFlow.fvs — Method

get ordered list of indices of dynamic variable for each flow

StockFlow.inflows — Method

return inflows of stock index s

StockFlow.inflowsAll — Method

return all inflows

StockFlow.instock — Method

return stocks of flow index f flow in

StockFlow.lpvvposition — Method

return argument position of source constant parameter of an auxiliary variable

StockFlow.lsnames — Method

return the pair of names of (stock, sum-auxiliary-variable) for all linkages between them

StockFlow.lsvvposition — Method

return argument position of source sum dynamical variable of an auxiliary variable

StockFlow.lvtgtposition — Method

return argument position of source auxiliary variable of an auxiliary variable

StockFlow.lvvposition — Method

return argument position of source stock variable of an auxiliary variable

StockFlow.make_v_expr — Method

create expresision of an auxiliary variable v

StockFlow.ne — Method

return count of edges of CLD

StockFlow.nf — Method

flow count

StockFlow.ni — Method

inflow count

StockFlow.nlpv — Method

link dynamic variable parameter count

StockFlow.nls — Method

link stock sum count

StockFlow.nlsv — Method

link sum variable dynamic variable count

StockFlow.nlv — Method

link stock dynamic variable count

StockFlow.nlvv — Method

link dynamic variable dynamic variable count

StockFlow.nn — Method

return count of nodes of CLD

StockFlow.nname — Method

return node's name with index n

StockFlow.nnames — Method

return node names of CLD

StockFlow.no — Method

outflow count

StockFlow.np — Method

parameter count

StockFlow.ns — Method

stock count

StockFlow.nsv — Method

sum variable count

StockFlow.nvb — Method

dynamic variable count

StockFlow.outflows — Method

return outflows of stock index s

StockFlow.outflowsAll — Method

return all outflows

StockFlow.outstock — Method

return stocks that flow index f flow out from

StockFlow.pname — Method

return the auxiliary variables name with index of v

StockFlow.pnames — Method

return parameter names

StockFlow.rebuildStratifiedModelByFlattenSymbols — Method

Return a new stock flow with flattened names, operators and positions from the old

StockFlow.sedge — Method

return edge's name with target number t

StockFlow.set_fnames! — Method

set flow names to vector of symbols

StockFlow.set_pnames! — Method

set parameter names to vector of symbols

StockFlow.set_snames! — Method

set stock names to vector of symbols

StockFlow.set_svnames! — Method

set sum variable names to vector of symbols

StockFlow.set_vnames! — Method

set dynamic variable names to vector of symbols

StockFlow.sname — Method

return the stocks name with index of s

StockFlow.snames — Method

return stock names

StockFlow.stockssv — Method

return stocks that sum variable index sv link to

StockFlow.stocksv — Method

return stocks that auxiliary variable index v link to

StockFlow.svname — Method

return the sum auxiliary variables name with index of sv

StockFlow.svnames — Method

return sum variable names

StockFlow.svsstock — Method

return sum auxiliary variables that stock s links to

StockFlow.svsstockAllF — Method

return sum auxiliary variables all stocks link (frequency)

StockFlow.svsv — Method

return sum variables that auxiliary variable index v link to

StockFlow.tedge — Method

return edge's name with edge number e

StockFlow.vname — Method

return the auxiliary variables name with index of v

StockFlow.vnames — Method

return variable names

StockFlow.vop — Method

return operator of an auxiliary variable

StockFlow.vpsrc — Method

return auxiliary variable's source constant parameters

StockFlow.vptgt — Method

return constant parameters's link auxiliary variables

StockFlow.vsrc — Method

return auxiliary variable's target auxiliary variables it links to

StockFlow.vsstock — Method

return auxiliary variables that stock s links to

StockFlow.vsstockAllF — Method

return auxiliary variables all stocks link (frequency)

StockFlow.vssv — Method

return auxiliary variables that sum auxiliary variable sv links to

StockFlow.vssvAllF — Method

return auxiliary variables all sum auxiliary variables link (frequency)

StockFlow.vtgt — Method

return auxiliary variable's source auxiliary variables that it links to

StockFlow.Syntax — Module

Alternative syntax for use in the definition of stock and flow models.

Examples

# An S-I-R model of infection
SIR = @stock_and_flow begin
    :stocks
    S
    I
    R

    :parameters
    c
    beta
    tRec

    :dynamic_variables
    v_prevalence = I / N
    v_meanInfectiousContactsPerS = c * v_prevalence
    v_perSIncidenceRate = beta * v_meanInfectiousContactsPerS
    v_newInfections = S * v_perSIncidenceRate
    v_newRecovery = I / tRec

    :flows
    S => inf(v_newInfections) => I
    I => rec(v_newRecovery) => R

    :sums
    N = [S, I, R]
end

# Generates:
# SIR = StockAndFlowF(
#     # stocks
#     (:S => (:F_NONE, :inf, :N), :I => (:inf, :rec, :N), :R => (:rec, :F_NONE, :N)),
#     # parameters
#     (:c, :beta, :tRec),
#     # dynamical variables
#     (   :v_prevalence => ((:I, :N) => :/),
#         :v_meanInfectiousContactsPerS => ((:c, :v_prevalence) => :*),
#         :v_perSIncidenceRate => ((:beta, :v_meanInfectiousContactsPerS) => :*),
#         :v_newInfections => ((:S, :v_perSIncidenceRate) => :*),
#         :v_newRecovery => ((:I, :tRec) => :/),
#     ),
#     # flows
#     (:inf => :v_newInfections, :rec => :v_newRecovery),
#     # sum dynamical variables
#     (:N),
# )

# The same model as before, but with the dynamic variables inferred
SIR_2 = @stock_and_flow begin
    :stocks
    S
    I
    R

    :parameters
    c
    beta
    tRec

    # We can leave out dynamic variables and let them be inferred from flows entirely!

    :flows
    S => inf(S * beta * (c * (I / N))) => I
    I => rec(I / tRec) => R

    :sums
    N = [S, I, R]
end

# Another possible S-I-R model definition
SIR_3 = @stock_and_flow begin
    :stocks
    S
    I
    R

    :parameters
    c
    beta
    tRec
    omega
    alpha

    :dynamic_variables
    v_prevalence = I / totalPopulation
    v_forceOfInfection = c * v_prevalence * beta

    :flows
    S => inf(S * v_forceOfInfection) => I
    ☁ => births(totalPopulation * alpha) => S
    S => deathsS(S * omega) => ☁
    I => rec(I / tRec) => R
    I => deathsI(I * omega) => ☁
    R => deathsR(R * omega) => ☁

    :sums
    totalPopulation = [S, I, R]
end

StockFlow.Syntax.infer_links — Method

infer_links(sfsrc :: StockAndFlowF, sftgt :: StockAndFlowF, NecMaps :: Dict{Symbol, Vector{Int64}})

Infer LS, I, O, LV, LSV, LVV, LPV mappings for an ACSetTransformation. Returns dictionary of Symbols to lists of indices, corresponding to an ACSetTransformation argument. If there exist no such mappings (eg, no LVV), that pairing will not be included in the returned dictionary.

If A <- C -> B, and we have A -> A' and B -> B' and a unique C' such that A' <- C' -> B', we can assume C -> C'.

:S => [2,4,1,3], :F => [1,2,4,3], ...

NecMaps must contain keys S, F, SV, P, V, each pointing to a (possibly empty) array of indices

StockFlow.Syntax.@feet — Macro

feet(block :: Expr)

Create Vector of feet using same notation for foot macro. Separated by newlines. First argument is stock, second is sum variable.

feetses = @feet begin
    A => B
    () => N
    C => ()
    D => E
    () => ()
    P => NI, R => NI, () => N
end

StockFlow.Syntax.@foot — Macro

foot(block :: Expr)

Create a foot with S => N syntax, where S is stock, N is sum variable.

@foot P => Q
@foot S1 => ()
@foot () => N
@foot () => ()
@foot A => N, () => NI

StockFlow.Syntax.@stock_and_flow — Macro

stock_and_flow(block :: Expr)

Compiles stock and flow syntax of the line-based block form

  :stocks
    symbol_1
    symbol_2
    ...
    symbol_n

  :parameters
    param_1
    param_2
    ...
    param_n

  :dynamic_variables
    dyvar_1 = symbol_h * param_g ... - symbol_x / param_y
    ...
    dyvar_n = symbol_k * param_j - dyvar_a ... - symbol_p / param_q

  :flows
    symbol_r => flow_name_1(dyvar_k) => symbol_q
    symbol_z => flow_name_2(dyvar_g * param_v) => symbol_p
    ☁       => flow_name_3(symbol_c + dyvar_b) => symbol_r
    symbol_j => flow_name_4(param_l + symbol_m) => CLOUD
    ...
    symbol_y => flow_name_n(dyvar_f) => ☁

into a StockAndFlowF data type for use with the StockFlow.jl modelling system.

StockFlow.PremadeModels.seir — Method

Return a new SEIR model

StockFlow.PremadeModels.sir — Method

Return a new SIR model

StockFlow.PremadeModels.sis — Method

Return a new SIS model

StockFlow.PremadeModels.svi — Method

Return a new SVI model